latest 0.5.11

+++ lib/features/cart/widgets/cart_subtotal.dart
@@ -8,10 +8,8 @@ class CartSubtotal extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final user = context.watch<UserProvider>().user;
-    int sum = 0;
-    user.cart
-        .map((e) => sum += e['quantity'] * e['product']['price'] as int)
-        .toList();
+    double sum = 0;
+    user.cart.map((e) => sum += e['quantity'] * e['product']['price']).toList();
 
     return Container(
       margin: const EdgeInsets.all(10),
